nexerra/inference/bio/pyproject.toml
[tool.poetry]
name = "phd-nexterrareward"
version = "0.1.0"
description = "Drug delivery reward function for MOF linker generation package"
authors = ["Ivan Zyuzin <iz233@cam.ac.uk>"]
license = "MIT"
readme = "readme.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = ">=2.1,<3.0"
numpy = ">=1.26,<2.0"
pyarrow = ">=14.0,<20.0"
rdkit-pypi = ">=2022.9.5"
jupyter = ">=1.1.1,<2.0.0"
ipykernel = ">=6.29,<7.0"
matplotlib = ">=3.8,<4.0"
python-dateutil = ">=2.8.2,<3.0"
tqdm = ">=4.66,<5.0"
umap-learn = "^0.5.11"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0,<9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
